[TASK] Fix the way the TypoScript configuration is fetched
Depending on whether the unit tests are launched from the `phpunit.xml` file,
or directly from within a unit test class, the relative path to the file
containing the default TypoScript configuration is not the same.

Here we fix this by setting the relative path from the PHP file which fetches
the configuration, so it always work.