[BUGFIX] Fix persistent option in field ViewHelper
When using the ViewHelper `<formz:option>` inside `<formz:field>`, the option
would not be deleted after the whole field is processed, resulting in unwanted
options in later fields, which could cause pretty ugly behaviours.

For instance, a field could be flagged as required even if it is not.

In the following example, the option `required` would have been present for the
field `bar` (it is now fixed by this commit).

```
<formz:field name="foo">
    <formz:option name="required" value="1" />

    [...]
</formz:field>

<formz:field name="bar">
    [...]
</formz:field>
```