[TASK] Disable escape output for some ViewHelpers
With the introduction of the standalone version of Fluid, the escaping handling
has been changed, meaning we don't have to use the `f:format.htmlentities` as
the content is escaped by default. Thus, we don't want some Formz ViewHelpers to
be escaped, so we need to fill the `$escapeOutput` property.

See [Feature-#69863](https://docs.typo3
.org/typo3cms/extensions/core/8-dev/Changelog/8
.0/Feature-69863-UseNewStandaloneFluidAsComposerDependency
.html#new-and-more-efficient-escaping)