[TASK] Change how `Message` instances are handled
A complete refactor has been done on the handling of the errors,
warnings and notices. Before, a ugly hack was done to determine the
validation name (for instance `required`) and the message key (for
instance `default`). The title of the message was used to store these
two values, separated by a `:`.

Now, a new abstraction class is introduced and used in the API, to
handle the whole system with a better approach.

The system still handles classic Extbase message instances, resulting in
a validation name and a message key both forced to `unknown`.