[BUGFIX] Fix page renderer singleton issue
The page renderer must be instantiated with `GeneralUtility::makeInstance()`
instead of Extbase object manager (or with an inject function).

This is due to some TYPO3 low level behaviour which overrides the page renderer
singleton instance, whenever a new request is used. The problem is that the
instance is not updated on Extbase side.

Using Extbase injection can lead to old page renderer instance being used,
resulting in a leak of assets inclusion, and maybe more issues.