[TASK] Improve assets inclusion to make it work in backend context
Because of a TYPO3 core issue, asset inclusion in the page footer wont work for
backend context. This commit will fix it by adding an abstraction function which
will handle both frontend and backend context. See this forge issue for more
information: https://forge.typo3.org/issues/60213

This commit also resolves #35: the default JavaScript files are now included in
the footer (in frontend context), as they already should have.

The JavaScript API changed a bit, it is more flexible to allow the registration
of a new form before the DOM has been fully loaded.