[TASK] Introduce new way of assets cache clearing
This commit changes the way the assets are cleared when the clear-caches
command is done.

Previously, the files would all be deleted, to be created once again on
the first request. This could lead to unwanted behaviours when using
load balanced servers. If a server did clear shared caches when another
server was dispatching a request (with `concatenate` configuration
enabled), the resulting merged file could be incomplete due to Formz
assets files missing. This wrong merged file would then be served at
each request.

Now, the files are not deleted when caches are cleared: their
modification time is forced to `0`, which allows current requests to
still be able to access to the file content.