[TASK] Migrate form-related features into `FormObject`
The `formWasSubmitted` and `formInstance` features were moved from the service
`FormViewHelperService` to `FormObject`. It makes more sense: it allows more
consistency in the whole code architecture as well as providing a new access to
this feature for some parts of the code.

A whole bunch of refactoring has been made in the entire code.