[FEATURE] Introduce `arguments` in `slot.render` view helper
You can now fill `arguments` with an array of arbitrary data that will be passed
to the slot and can be used within it.

Note that if an argument defined in this list is also filled in the
``arguments`` of the slot, it will be overridden with the value defined in the
slot definition.

**Example:**

```html
<formz:slot.render slot="My.Slot" arguments="{class: 'my-class'}" />
```