[BUGFIX] Fix cHash issue in Ajax request URI
Adding an argument forces the URI to calculate a cHash (it was not the case
previously). Unfortunately, in TYPO3 8 the cHash is required in the request, and
would cause a fatal error when it's missing.

This may be a temporary fix, as this is probably not the best way to do it.