[!!!][TASK] Rename `form configuration` to `form definition`
The whole configuration objects representing all form properties have been moved
and renamed to `definition`, as this word better represents what the objects
actually are and do.

This commit is part of a bigger plan to refactor the form objects handling, that
will result in an easier way to manipulate the forms during the runtime.