[!!!][FEATURE] Improve form object handling
The form object handling has been been completely refactored.

This will ease form manipulation: developers will be able to access most of the
data that is directly bound to a form.

A form object is divided in two objects (which are invisible for the developer):

 - A static object, that contains the form definition - which is by default
   handled by TypoScript - and potential configuration validation errors.
 - A proxy object, which is directly bound to the form instance (the class that
   implements `FormInterface`).

The form object is a bridge between these two objects, and contains almost only
getter functions.

This new feature will be used for incoming new features (middlewares and more).