[!!!][TASK] Change condition configuration validation implementation
The validation function for a condition configuration has been changed: the
function now has a parameter of type `FormDefinition` which must be used to
check for possible configuration error(s).

The internal properties of the class must **NOT** be used anymore, as they may
not have been bound to the instance when the validation function is called.
Instead, use the new function parameter.