[FEATURE] Introduce post configuration process signal
This signal can be used to modify the configuration object after it
has been built using the TypoScript configuration.

Note that the signal will be dispatched only once, as the configuration
object is latter put in cache. Therefore, your slots will be called only
once too and need to be aware of it.

Example (`ext_localconf.php`):

```
$signalSlotDispatcher =
    \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(
        \TYPO3\CMS\Extbase\SignalSlot\Dispatcher::class
    );
$signalSlotDispatcher->connect(
    \Romm\Formz\Configuration\ConfigurationFactory::class,
    \Romm\Formz\Configuration\ConfigurationFactory::POST_CONFIGURATION_PROCESS_SIGNAL,
    function (\Romm\Formz\Configuration\Configuration $configuration) {
        $configuration->getView()->setPartialRootPath(
            50,
            'EXT:my_extension/Resources/Private/Partials/FormZ/'
        );
    }
);
```