[FEATURE] Improve form definition manual manipulation
This commit introduces a major refactoring of both the root configuration and
the form definition, by making them a lot easier to manipulate (with the
configuration post process signal and the form object builder).

A "freeze" state is also introduced, meaning configuration/definition cannot be
manipulated anymore when the builders turn the freeze state on. This is done to
prevent the configuration to be edited during runtime.