[BUGFIX] Fix layout/partial root paths not merged in field layouts
This commit fixes the situation where a slot in a field layout tries to use a
partial from the actual rendering context: until now only the partials
configured in the FormZ view TypoScript configuration were supported, resulting
in a fatal error.

The paths are now merged together, giving access to both of them.

The same behaviour is done for layouts.