[FEATURE] Introduce `getProperties` method in form object
This method returns an array containing the names of the properties that can be
found in the form class.

These properties should be directly bound to the form: properties containing
other values like a service should not be in this list. To prevent it, the tag
`@formz-ignore` must be added in the property annotation.

Example:

```
class MyForm implements FormInterface
{
    /**
     * @var string
     */
    protected $email;

    /**
     * @var SomeService
     * @formz-ignore
     */
    protected $service
}
```